# models.py
import json
from pydantic import BaseModel
import typing


class Coords(BaseModel):
    latitude: float
    longitude: float


class Corner(BaseModel):
    x: int
    y: int


class Version(BaseModel):
    major: int = 0
    minor: int = 0
    patch: int = 1

    def __str__(self):
        return str(self.major)+"."+str(self.minor)+"."+str(self.patch)


class QAItem(BaseModel):
    question: str
    answer: str


class ChecklistItem(BaseModel):
    description: str
    checked: bool = False


class BalconyImageIn(BaseModel):
    base64: str
    img: typing.Any | None = None
    width: int | None = 0
    height: int | None = 0
    url: str | None = None


class BalconyImageOut(BaseModel):
    area: int = 0
    boundary: tuple[int, int, int, int] = (0, 0, 0, 0)
    corners: list[Corner] = []


class UserDataBalcony(BaseModel):
    alignment: str = 'S'
    shadowing: float = 1.0


class UserDataPV(BaseModel):
    module_count: int = 2
    module_power: int = 300
    angle: int = 90


class UserDataConsumption(BaseModel):
    amount: int
    price: float


class UserDataIn(BaseModel):
    DataProcessingAccepted: bool = False
    Location: Coords
    PV: UserDataPV
    Balcony: UserDataBalcony
    Consumption: UserDataConsumption
    TimePeriod: int = 1


class KpiResult(BaseModel):
    energy_output_per_year: float = 0.0
    amortization: float = 0.0
    savings: float = 0.0
    self_sufficiency: float = 0.0
    self_consumption: float = 0.0


class MastrDataOut(BaseModel):
    id: str = ''
    name: str = ''
    state: str | None = None
    zipCode: int | None = None
    street: str | None = None

    def __str__(self):
        return f'{self.name} ({self.id}) - {self.street} {self.zipCode} {self.state}'