# openapi.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "FastAPI",
        "version": "0.1.0"
    },
    "paths": {
        "/api": {
            "get": {
                "summary": "Version",
                "operationId": "version_api_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Version"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/coords": {
            "get": {
                "summary": "Coords",
                "operationId": "coords_api_coords_get",
                "parameters": [
                    {
                        "required": true,
                        "schema": {
                            "title": "Zip",
                            "type": "integer"
                        },
                        "name": "zip",
                        "in": "query"
                    },
                    {
                        "required": true,
                        "schema": {
                            "title": "City",
                            "type": "string"
                        },
                        "name": "city",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Coords"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/info": {
            "get": {
                "summary": "Info",
                "operationId": "info_api_info_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "title": "Response Info Api Info Get",
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/QA"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/balcony": {
            "post": {
                "summary": "Balcony",
                "operationId": "balcony_api_balcony_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BalconyImageIn"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BalconyImageOut"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/kpi": {
            "post": {
                "summary": "Kpi",
                "operationId": "kpi_api_kpi_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserDataIn"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/KpiResult"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/ar": {
            "get": {
                "summary": "Ar",
                "operationId": "ar_api_ar_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/checklist": {
            "get": {
                "summary": "Checklist",
                "operationId": "checklist_api_checklist_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "title": "Response Checklist Api Checklist Get",
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ChecklistItem"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/mastr": {
            "get": {
                "summary": "Mastr",
                "operationId": "mastr_api_mastr_get",
                "parameters": [
                    {
                        "required": true,
                        "schema": {
                            "title": "Q",
                            "type": "string"
                        },
                        "name": "q",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "title": "Response Mastr Api Mastr Get",
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/MastrDataOut"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "BalconyImageIn": {
                "title": "BalconyImageIn",
                "required": [
                    "base64"
                ],
                "type": "object",
                "properties": {
                    "base64": {
                        "title": "Base64",
                        "type": "string"
                    },
                    "img": {
                        "title": "Img"
                    },
                    "width": {
                        "title": "Width",
                        "type": "integer",
                        "default": 0
                    },
                    "height": {
                        "title": "Height",
                        "type": "integer",
                        "default": 0
                    },
                    "url": {
                        "title": "Url",
                        "type": "string"
                    }
                }
            },
            "BalconyImageOut": {
                "title": "BalconyImageOut",
                "type": "object",
                "properties": {
                    "area": {
                        "title": "Area",
                        "type": "integer",
                        "default": 0
                    },
                    "boundary": {
                        "title": "Boundary",
                        "maxItems": 4,
                        "minItems": 4,
                        "type": "array",
                        "items": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "integer"
                            },
                            {
                                "type": "integer"
                            },
                            {
                                "type": "integer"
                            }
                        ],
                        "default": [
                            0,
                            0,
                            0,
                            0
                        ]
                    },
                    "corners": {
                        "title": "Corners",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Corner"
                        },
                        "default": []
                    }
                }
            },
            "ChecklistItem": {
                "title": "ChecklistItem",
                "required": [
                    "description"
                ],
                "type": "object",
                "properties": {
                    "description": {
                        "title": "Description",
                        "type": "string"
                    },
                    "checked": {
                        "title": "Checked",
                        "type": "boolean",
                        "default": false
                    }
                }
            },
            "Coords": {
                "title": "Coords",
                "required": [
                    "latitude",
                    "longitude"
                ],
                "type": "object",
                "properties": {
                    "latitude": {
                        "title": "Latitude",
                        "type": "number"
                    },
                    "longitude": {
                        "title": "Longitude",
                        "type": "number"
                    }
                }
            },
            "Corner": {
                "title": "Corner",
                "required": [
                    "x",
                    "y"
                ],
                "type": "object",
                "properties": {
                    "x": {
                        "title": "X",
                        "type": "integer"
                    },
                    "y": {
                        "title": "Y",
                        "type": "integer"
                    }
                }
            },
            "HTTPValidationError": {
                "title": "HTTPValidationError",
                "type": "object",
                "properties": {
                    "detail": {
                        "title": "Detail",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        }
                    }
                }
            },
            "KpiResult": {
                "title": "KpiResult",
                "type": "object",
                "properties": {
                    "energy_output_per_year": {
                        "title": "Energy Output Per Year",
                        "type": "number",
                        "default": 0.0
                    },
                    "amortization": {
                        "title": "Amortization",
                        "type": "number",
                        "default": 0.0
                    },
                    "savings": {
                        "title": "Savings",
                        "type": "number",
                        "default": 0.0
                    },
                    "self_sufficiency": {
                        "title": "Self Sufficiency",
                        "type": "number",
                        "default": 0.0
                    },
                    "self_consumption": {
                        "title": "Self Consumption",
                        "type": "number",
                        "default": 0.0
                    }
                }
            },
            "MastrDataOut": {
                "title": "MastrDataOut",
                "type": "object",
                "properties": {
                    "id": {
                        "title": "Id",
                        "type": "string",
                        "default": ""
                    },
                    "name": {
                        "title": "Name",
                        "type": "string",
                        "default": ""
                    },
                    "state": {
                        "title": "State",
                        "type": "string"
                    },
                    "zipCode": {
                        "title": "Zipcode",
                        "type": "integer"
                    },
                    "street": {
                        "title": "Street",
                        "type": "string"
                    }
                }
            },
            "QA": {
                "title": "QA",
                "required": [
                    "question",
                    "answer"
                ],
                "type": "object",
                "properties": {
                    "question": {
                        "title": "Question",
                        "type": "string"
                    },
                    "answer": {
                        "title": "Answer",
                        "type": "string"
                    }
                }
            },
            "UserDataBalcony": {
                "title": "UserDataBalcony",
                "type": "object",
                "properties": {
                    "alignment": {
                        "title": "Alignment",
                        "type": "string",
                        "default": "S"
                    },
                    "shadowing": {
                        "title": "Shadowing",
                        "type": "number",
                        "default": 1.0
                    }
                }
            },
            "UserDataConsumption": {
                "title": "UserDataConsumption",
                "required": [
                    "amount",
                    "price"
                ],
                "type": "object",
                "properties": {
                    "amount": {
                        "title": "Amount",
                        "type": "integer"
                    },
                    "price": {
                        "title": "Price",
                        "type": "number"
                    }
                }
            },
            "UserDataIn": {
                "title": "UserDataIn",
                "required": [
                    "Location",
                    "PV",
                    "Balcony",
                    "Consumption"
                ],
                "type": "object",
                "properties": {
                    "DataProcessingAccepted": {
                        "title": "Dataprocessingaccepted",
                        "type": "boolean",
                        "default": false
                    },
                    "Location": {
                        "$ref": "#/components/schemas/Coords"
                    },
                    "PV": {
                        "$ref": "#/components/schemas/UserDataPV"
                    },
                    "Balcony": {
                        "$ref": "#/components/schemas/UserDataBalcony"
                    },
                    "Consumption": {
                        "$ref": "#/components/schemas/UserDataConsumption"
                    },
                    "TimePeriod": {
                        "title": "Timeperiod",
                        "type": "integer",
                        "default": 1
                    }
                }
            },
            "UserDataPV": {
                "title": "UserDataPV",
                "type": "object",
                "properties": {
                    "module_count": {
                        "title": "Module Count",
                        "type": "integer",
                        "default": 2
                    },
                    "module_power": {
                        "title": "Module Power",
                        "type": "integer",
                        "default": 300
                    },
                    "angle": {
                        "title": "Angle",
                        "type": "integer",
                        "default": 90
                    }
                }
            },
            "ValidationError": {
                "title": "ValidationError",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "type": "object",
                "properties": {
                    "loc": {
                        "title": "Location",
                        "type": "array",
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        }
                    },
                    "msg": {
                        "title": "Message",
                        "type": "string"
                    },
                    "type": {
                        "title": "Error Type",
                        "type": "string"
                    }
                }
            },
            "Version": {
                "title": "Version",
                "type": "object",
                "properties": {
                    "major": {
                        "title": "Major",
                        "type": "integer",
                        "default": 0
                    },
                    "minor": {
                        "title": "Minor",
                        "type": "integer",
                        "default": 0
                    },
                    "patch": {
                        "title": "Patch",
                        "type": "integer",
                        "default": 1
                    }
                }
            }
        }
    }
}